refactor(exoplayer): extract periodic time elapsed handler
