+++ player/src/main/kotlin/com/globo/clappr/playback/ExoPlayerPlayBack.kt
@@ -216,40 +216,6 @@ open class ExoPlayerPlayback(source: String, mimeType: String? = null, options: 
     }
 
 
-    class PeriodicTimeElapsedHandler(val interval: Long, val function: () -> Unit) : Handler() {
-
-        private var timeElapsedRunnable: TimeElapsedRunnable? = null
-
-        fun start() {
-            timeElapsedRunnable?.cancel()
-            timeElapsedRunnable = TimeElapsedRunnable(this)
-            postDelayed(timeElapsedRunnable, interval)
-        }
-
-        fun cancel() {
-            timeElapsedRunnable?.cancel()
-            removeCallbacks(timeElapsedRunnable)
-            timeElapsedRunnable = null
-        }
-
-        inner class TimeElapsedRunnable(val handler: Handler) : Runnable {
-
-            private var canceled: Boolean = false
-
-            override fun run() {
-
-                if (!canceled) {
-                    function.invoke()
-                    handler.postDelayed(this, interval)
-                }
-            }
-
-            fun cancel() {
-                canceled = true
-            }
-        }
-    }
-
     inner class ExoplayerEventsListener() : AdaptiveMediaSourceEventListener, ExtractorMediaSource.EventListener, ExoPlayer.EventListener {
         override fun onLoadError(error: IOException?) {
             handleError(error)
